requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# LLM
openai==1.10.0

# Embeddings & Vector Store
sentence-transformers==2.3.1
chromadb==0.4.22
pinecone-client

# Scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Utilities
tqdm==4.66.1
